fix: Add type assertion for TextDecoder in CERI AI stream

+++ src/app/api/ceri-ai/stream/route.ts
@@ -92,7 +92,7 @@ export async function POST(req: NextRequest) {
             // Accumulate response for caching
             if (useCache && preferUpstash() && upstashCache.isEnabled()) {
               try {
-                const chunkText = new TextDecoder().decode(chunk)
+                const chunkText = new TextDecoder().decode(chunk as Uint8Array)
                 const lines = chunkText.split('\n')
                 for (const line of lines) {
                   if (line.startsWith('data: ')) {
